requirements.txt
# Core Django framework
# Updated: 2025-01-14 - Removed consul and other Python 3.13 incompatible packages
Django
django-cors-headers
django-environ

# Django REST Framework
djangorestframework

# Database
psycopg2-binary
dj-database-url

# Web server and static files
gunicorn
whitenoise

# Environment and configuration
python-dotenv

# Authentication and security
PyJWT
cryptography

# HTTP client
requests
httpx

# Data processing
pandas
numpy

# Redis for caching
redis
hiredis

# Background tasks
celery

# Google Calendar Integration
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Email
sendgrid

# Push notifications
pywebpush
py-vapid

# Utilities
python-dateutil
pytz

# Development and testing
pytest
pytest-django

# API framework (FastAPI for microservices)
fastapi
uvicorn
pydantic

# Async support
aiofiles

# JSON handling
orjson

# Validation
email-validator

# File handling
Pillow

# Monitoring and system info
structlog
psutil

# Time utilities
arrow

# Supabase client
supabase

# AWS SDK (for backup/storage)
boto3

# Web push notifications (webpush-data-encryption removed - no Python 3.13 compatible version)
# Using pywebpush and py-vapid instead for web push functionality
